refactor: extract types/proxy/error/fixtures from client_events.rs
client_events.rs was 2692 lines containing unrelated cohesive chunks.
Extracted into the existing client_events/ submodule directory so each
concern lives in its own file, leaving the dangerous process_open_request
dispatcher untouched for a follow-up.

- types.rs: RequestId, ClientId, AuthToken, OpenRequest, HostResult,
  HostIncomingMsg + their thread-local counters (verbatim)
- proxy.rs: ClientEventsProxy trait + BoxedClient type alias (verbatim)
- error.rs: Error enum + From<RingError> + ensure_peer_ready (verbatim)
- test.rs: pub(crate) mod test body (MemoryEventsGen, NetworkEventGenerator,
  RandomEventGenerator, InternalGeneratorState) (verbatim)

Public API is fully preserved via re-exports in client_events.rs; all
existing use crate::client_events::X imports continue to resolve.
